+++ src/core/components/welcome.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Modal from "./modal";
+import Modal from "../ui/modal";
 
 export default class Welcome extends React.Component {
